backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-jose[cryptography]
bcrypt
python-dotenv
requests
psycopg2-binary
